Pay Breakdown
This is a super simple function that I made to save me a little bit of time when I get my paycheck. I used to sit and make these calculations one by one every time I got a paycheck, so I wrote this quick function to do it for me. Just input the amount of my take home pay and then the calculations are done. The 'xxx' represents static amounts that don't change, for example my bills (cable, internet, netflix etc.) and the minimum monthly payment of my student loans. The only minorly tricky thing for me to figure out was how to get Python to cut off after two decimal places. I found the "%.2f" solution on stack overflow and it works perfectly. 

I'm planning on adding to this as I start to learn more. I'd like to add some functionality that would allow me to save and recall past amounts from past pay periods, and then be able to visually represent some of the info. This is the first step though.